Updated example naming
